+++ src/resampler.rs
@@ -134,6 +134,16 @@ pub struct Resampler<
     input_start: Option<DateTime<Utc>>,
     /// The interval between the first and the second sample in the buffer
     input_interval: Option<TimeDelta>,
+    /// Whether the resampled timestamp should be the first timestamp (if
+    /// `first_timestamp` is `true`) or the last timestamp (if
+    /// `first_timestamp` is `false`) in the buffer.
+    /// If `first_timestamp` is `true`, the resampled timestamp will be the
+    /// timestamp of the first sample in the buffer and the aggregation will
+    /// be done with the samples that are `interval` in the future.
+    /// If `first_timestamp` is `false`, the resampled timestamp will be the
+    /// timestamp of the last sample in the buffer and the aggregation will
+    /// be done with the samples that are `interval` in the past.
+    first_timestamp: bool,
 }
 
 impl<
@@ -148,6 +158,7 @@ impl<
         resampling_function: ResamplingFunction<T, S>,
         max_age_in_intervals: i32,
         start: DateTime<Utc>,
+        first_timestamp: bool,
     ) -> Self {
         let aligned_start = epoch_align(interval, start, None);
         Self {
@@ -155,6 +166,7 @@ impl<
             resampling_function,
             max_age_in_intervals,
             start: aligned_start,
+            first_timestamp,
             ..Default::default()
         }
     }
@@ -181,12 +193,23 @@ impl<
         let mut buffer_iter = self.buffer.iter();
         let mut next_sample: Option<&S> = buffer_iter.next();
         self.input_start = next_sample.map(|s| s.timestamp());
+        let offset = if self.first_timestamp {
+            TimeDelta::zero()
+        } else {
+            self.interval
+        };
 
         // loop over the intervals
         while self.start < end {
             // loop over the samples in the buffer
             while next_sample
-                .map(|s| s.timestamp() < self.start + self.interval)
+                .map(|s| {
+                    is_left_of_buffer_edge(
+                        self.first_timestamp,
+                        &s.timestamp(),
+                        &(self.start + self.interval),
+                    )
+                })
                 .unwrap_or(false)
             {
                 // next sample is not newer than the current interval
@@ -211,11 +234,13 @@ impl<
             let input_interval = self.input_interval.unwrap_or(self.interval);
             let drain_end_date =
                 self.start + self.interval - input_interval * self.max_age_in_intervals;
-            interval_buffer.retain(|s| s.timestamp() >= drain_end_date);
+            interval_buffer.retain(|s| {
+                is_right_of_buffer_edge(self.first_timestamp, &s.timestamp(), &drain_end_date)
+            });
 
             // resample the interval_buffer
             res.push(Sample::new(
-                self.start + self.interval,
+                self.start + offset,
                 self.resampling_function.apply(interval_buffer.as_slice()),
             ));
 
@@ -226,7 +251,9 @@ impl<
         // Remove samples from buffer that are older than max_age
         let interval = self.input_interval.unwrap_or(self.interval);
         let drain_end_date = end - interval * self.max_age_in_intervals;
-        self.buffer.retain(|s| s.timestamp() >= drain_end_date);
+        self.buffer.retain(|s| {
+            is_right_of_buffer_edge(self.first_timestamp, &s.timestamp(), &drain_end_date)
+        });
 
         res
     }
@@ -260,4 +287,28 @@ pub(crate) fn epoch_align(
             + alignment_timestamp.timestamp_millis(),
     )
     .unwrap_or(timestamp)
+}
+
+fn is_left_of_buffer_edge(
+    first_timestamp: bool,
+    timestamp: &DateTime<Utc>,
+    edge_timestamp: &DateTime<Utc>,
+) -> bool {
+    if first_timestamp {
+        timestamp < edge_timestamp
+    } else {
+        timestamp <= edge_timestamp
+    }
+}
+
+fn is_right_of_buffer_edge(
+    first_timestamp: bool,
+    timestamp: &DateTime<Utc>,
+    edge_timestamp: &DateTime<Utc>,
+) -> bool {
+    if first_timestamp {
+        timestamp >= edge_timestamp
+    } else {
+        timestamp > edge_timestamp
+    }
 }